type : 타입 수정
